Fail CI if go tests fail in a nix sandbox (#134)
* remove subPackages from package.nix so that all tests are run

`subPackages`` has a bug that if it is set, then only tests for
`cmd/paretosecurity`` are run, which effectively means no tests are
run. There's a draft PR to fix this, but it's not gonna happen soon,
so let's just not use `subPackages`` for now.

* refactor: move checkPortMock to shared package for better testability

* fix: ensure config is reset after tests and prevent saving during testing

* test: add mock response for successful link command execution

* test: add unit tests for JWT parsing and link command execution

* test: add unit tests for JWT parsing and link command execution
refactor: remove config saving during tests to improve test isolation

* test: increase timeout for broadcast message reception in tests

* fix: change fatal log to warning for home directory retrieval failure

---------

Co-authored-by: Janez T <hey@dz0ny.dev>